Add privateroute and fix some path
