+++ src/context/AuthContext.jsx
@@ -1,5 +1,5 @@
 import { createContext, useEffect, useState, useContext } from "react";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../utils/supabase";
 const AuthContext = createContext();
 
 export const AuthContextProvider = ({ children }) => {
