+++ src/pages/UpdatePassword.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import { Eye, EyeOff, Lock } from "lucide-react";
 import { UserAuth } from "../context/AuthContext";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../utils/supabase";
 
 const UpdatePassword = () => {
     const [newPassword, setNewPassword] = useState("");
