--- src/router.jsx
@@ -1,18 +1,0 @@
-import { createBrowserRouter } from "react-router-dom";
-import Home from "./pages/Home";
-import Login from "./pages/Login";
-import Signup from "./pages/Signup";
-import NurseDashboard from "./pages/NurseDashboard";
-import ForgotPassword from "./pages/ForgotPassword";
-import UpdatePassword from "./pages/UpdatePassword";
-
-const router = createBrowserRouter([
-  { path: "/", element: <Home /> },
-  { path: "/login", element: <Login /> },
-  { path: "/signup", element: <Signup /> },
-  { path: "/forgot-password", element: <ForgotPassword /> },
-  { path: "/update-password", element: <UpdatePassword /> },
-  { path: "/nurse", element: <NurseDashboard />}
-]);
-
-export default router;
