+++ src/utils/privaterouter.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { UserAuth } from "../context/AuthContext";
+import { Navigate } from "react-router-dom";
+const PrivateRouter = ({children}) => {
+    const { session } = UserAuth();
+    if (session === undefined) {
+        return <p>Loading...</p>
+    }
+    return session ? children : <Navigate to="/login" replace/>
+}
+
+export default PrivateRouter;
