+++ src/utils/supabase.js
@@ -2,4 +2,5 @@ import {createClient} from '@supabase/supabase-js';
 
 const supabaseUrl = import.meta.env.VITE_SUPABASE_URL;
 const supabaseAnonKey = import.meta.env.VITE_SUPABASE_ANON_KEY;
+
 export const supabase = createClient(supabaseUrl, supabaseAnonKey);
